fix-ci: share comparison-method visitor to pass the quality gate
SonarQube Code Analysis failed on 12.7% duplication with S9148 and 89.6%
coverage on new code. Extract the shared Comparable/Comparator dispatch
and add a block-lambda case so S9354 is no longer a near-copy of S9148.
