+++ java-checks/src/main/java/org/sonar/java/checks/FloatingPointComparisonCheck.java
@@ -16,58 +16,27 @@
  */
 package org.sonar.java.checks;
 
-import java.util.Arrays;
-import java.util.List;
 import org.sonar.check.Rule;
-import org.sonar.plugins.java.api.IssuableSubscriptionVisitor;
-import org.sonar.plugins.java.api.semantic.MethodMatchers;
 import org.sonar.plugins.java.api.semantic.Type;
-import org.sonar.plugins.java.api.tree.BaseTreeVisitor;
 import org.sonar.plugins.java.api.tree.BinaryExpressionTree;
-import org.sonar.plugins.java.api.tree.ClassTree;
 import org.sonar.plugins.java.api.tree.ExpressionTree;
 import org.sonar.plugins.java.api.tree.LambdaExpressionTree;
 import org.sonar.plugins.java.api.tree.MethodTree;
 import org.sonar.plugins.java.api.tree.Tree;
 
 @Rule(key = "S9148")
-public class FloatingPointComparisonCheck extends IssuableSubscriptionVisitor {
+public class FloatingPointComparisonCheck extends AbstractComparisonMethodCheck {
 
   private static final String MESSAGE = "Use \"Double.compare\" or \"Float.compare\" to compare floating-point values.";
 
-  private static final MethodMatchers COMPARE_METHODS = MethodMatchers.or(
-    MethodMatchers.create()
-      .ofSubTypes("java.lang.Comparable")
-      .names("compareTo")
-      .addParametersMatcher(MethodMatchers.ANY)
-      .build(),
-    MethodMatchers.create()
-      .ofSubTypes("java.util.Comparator")
-      .names("compare")
-      .addParametersMatcher(MethodMatchers.ANY, MethodMatchers.ANY)
-      .build());
-
   @Override
-  public List<Tree.Kind> nodesToVisit() {
-    return Arrays.asList(Tree.Kind.METHOD, Tree.Kind.LAMBDA_EXPRESSION);
+  void visitComparisonMethod(MethodTree methodTree) {
+    methodTree.block().accept(new FloatingPointComparisonVisitor());
   }
 
   @Override
-  public void visitNode(Tree tree) {
-    if (context.getSemanticModel() == null) {
-      return;
-    }
-    if (tree.is(Tree.Kind.METHOD)) {
-      MethodTree methodTree = (MethodTree) tree;
-      if (COMPARE_METHODS.matches(methodTree) && methodTree.block() != null) {
-        methodTree.block().accept(new FloatingPointComparisonVisitor());
-      }
-    } else {
-      LambdaExpressionTree lambda = (LambdaExpressionTree) tree;
-      if (lambda.symbolType().isSubtypeOf("java.util.Comparator")) {
-        lambda.body().accept(new FloatingPointComparisonVisitor());
-      }
-    }
+  void visitComparatorLambda(LambdaExpressionTree lambda) {
+    lambda.body().accept(new FloatingPointComparisonVisitor());
   }
 
   private static boolean hasFloatingType(ExpressionTree tree) {
@@ -75,7 +44,7 @@ public class FloatingPointComparisonCheck extends IssuableSubscriptionVisitor {
       || tree.symbolType().isPrimitive(Type.Primitives.DOUBLE);
   }
 
-  private class FloatingPointComparisonVisitor extends BaseTreeVisitor {
+  private class FloatingPointComparisonVisitor extends IgnoreNestedTypesVisitor {
 
     @Override
     public void visitBinaryExpression(BinaryExpressionTree tree) {
@@ -89,16 +58,6 @@ public class FloatingPointComparisonCheck extends IssuableSubscriptionVisitor {
 
     private boolean hasFloatingOperand(BinaryExpressionTree tree) {
       return hasFloatingType(tree.leftOperand()) || hasFloatingType(tree.rightOperand());
-    }
-
-    @Override
-    public void visitClass(ClassTree tree) {
-      // Do not visit inner classes
-    }
-
-    @Override
-    public void visitLambdaExpression(LambdaExpressionTree tree) {
-      // Do not visit nested lambdas
     }
   }
 }
