+++ java-checks/src/main/java/org/sonar/java/checks/IntegerSubtractionInComparisonCheck.java
@@ -16,16 +16,10 @@
  */
 package org.sonar.java.checks;
 
-import java.util.Arrays;
-import java.util.List;
 import org.sonar.check.Rule;
 import org.sonar.java.model.ExpressionUtils;
-import org.sonar.plugins.java.api.IssuableSubscriptionVisitor;
-import org.sonar.plugins.java.api.semantic.MethodMatchers;
 import org.sonar.plugins.java.api.semantic.Type;
-import org.sonar.plugins.java.api.tree.BaseTreeVisitor;
 import org.sonar.plugins.java.api.tree.BinaryExpressionTree;
-import org.sonar.plugins.java.api.tree.ClassTree;
 import org.sonar.plugins.java.api.tree.ExpressionTree;
 import org.sonar.plugins.java.api.tree.LambdaExpressionTree;
 import org.sonar.plugins.java.api.tree.MethodTree;
@@ -34,52 +28,27 @@ import org.sonar.plugins.java.api.tree.Tree;
 import org.sonar.plugins.java.api.tree.TypeCastTree;
 
 @Rule(key = "S9354")
-public class IntegerSubtractionInComparisonCheck extends IssuableSubscriptionVisitor {
+public class IntegerSubtractionInComparisonCheck extends AbstractComparisonMethodCheck {
 
   private static final String MESSAGE = "Subtracting numeric values in %s can overflow; use %s instead.";
 
-  private static final MethodMatchers COMPARE_METHODS = MethodMatchers.or(
-    MethodMatchers.create()
-      .ofSubTypes("java.lang.Comparable")
-      .names("compareTo")
-      .addParametersMatcher(MethodMatchers.ANY)
-      .build(),
-    MethodMatchers.create()
-      .ofSubTypes("java.util.Comparator")
-      .names("compare")
-      .addParametersMatcher(MethodMatchers.ANY, MethodMatchers.ANY)
-      .build());
-
   @Override
-  public List<Tree.Kind> nodesToVisit() {
-    return Arrays.asList(Tree.Kind.METHOD, Tree.Kind.LAMBDA_EXPRESSION);
+  void visitComparisonMethod(MethodTree methodTree) {
+    methodTree.block().accept(new ComparisonResultVisitor(methodTree.simpleName().name()));
   }
 
   @Override
-  public void visitNode(Tree tree) {
-    if (context.getSemanticModel() == null) {
-      return;
-    }
-    if (tree.is(Tree.Kind.METHOD)) {
-      MethodTree methodTree = (MethodTree) tree;
-      if (COMPARE_METHODS.matches(methodTree) && methodTree.block() != null) {
-        methodTree.block().accept(new ComparisonResultVisitor(methodTree.simpleName().name()));
-      }
+  void visitComparatorLambda(LambdaExpressionTree lambda) {
+    Tree body = lambda.body();
+    ComparisonResultVisitor visitor = new ComparisonResultVisitor("compare");
+    if (body.is(Tree.Kind.BLOCK)) {
+      body.accept(visitor);
     } else {
-      LambdaExpressionTree lambda = (LambdaExpressionTree) tree;
-      if (lambda.symbolType().isSubtypeOf("java.util.Comparator")) {
-        Tree body = lambda.body();
-        ComparisonResultVisitor visitor = new ComparisonResultVisitor("compare");
-        if (body.is(Tree.Kind.BLOCK)) {
-          body.accept(visitor);
-        } else {
-          visitor.checkComparisonResult((ExpressionTree) body);
-        }
-      }
+      visitor.checkComparisonResult((ExpressionTree) body);
     }
   }
 
-  private class ComparisonResultVisitor extends BaseTreeVisitor {
+  private class ComparisonResultVisitor extends IgnoreNestedTypesVisitor {
 
     private final String enclosingMethodName;
 
@@ -104,16 +73,6 @@ public class IntegerSubtractionInComparisonCheck extends IssuableSubscriptionVis
       if (replacement != null) {
         reportIssue(((BinaryExpressionTree) unwrapped).operatorToken(), String.format(MESSAGE, enclosingMethodName, replacement));
       }
-    }
-
-    @Override
-    public void visitClass(ClassTree tree) {
-      // Do not visit inner classes
-    }
-
-    @Override
-    public void visitLambdaExpression(LambdaExpressionTree tree) {
-      // Do not visit nested lambdas
     }
   }
